+++ peregrine-web/src/webgl/program/texture.rs
@@ -1,17 +1,23 @@
 use anyhow::{ anyhow as err };
 use web_sys::{ WebGlRenderingContext, WebGlTexture, HtmlCanvasElement };
-use super::values::ProcessValueType;
 use crate::webgl::canvas::canvas::{ Canvas, CanvasWeave };
 use crate::webgl::util::handle_context_errors;
 
-fn create_texture(context: &WebGlRenderingContext, element: &HtmlCanvasElement) -> anyhow::Result<()> {
+
+fn create_texture(context: &WebGlRenderingContext, our_data: (u32,Canvas)) -> anyhow::Result<(u32,WebGlTexture)> {
+    let texture = context.create_texture().ok_or_else(|| err!("cannot create texture"))?;
+    handle_context_errors(context)?;
+    context.bind_texture(WebGlRenderingContext::TEXTURE_2D,Some(&texture));
+    handle_context_errors(context)?;
     context.tex_image_2d_with_u32_and_u32_and_canvas( // wow
         WebGlRenderingContext::TEXTURE_2D,0,WebGlRenderingContext::RGBA as i32,WebGlRenderingContext::RGBA,
-        WebGlRenderingContext::UNSIGNED_BYTE,element
+        WebGlRenderingContext::UNSIGNED_BYTE,our_data.1.element()
     );
     handle_context_errors(context)?;
-    Ok(())
+    apply_weave(context,our_data.1.weave())?;
+    Ok((our_data.0,texture))
 }
+
 
 fn apply_weave(context: &WebGlRenderingContext,weave: &CanvasWeave) -> anyhow::Result<()> {
     let (minf,magf,wraps,wrapt) = match weave {
@@ -50,33 +56,33 @@ impl Texture {
     }
 }
 
-impl ProcessValueType for Texture {
-    type OurValue = (u32,Canvas);
-    type GLValue = (u32,WebGlTexture);
+pub(crate) struct TextureValues {
+    gl_value: Option<(u32,WebGlTexture)>,
+    object: Texture
+}
 
-    fn name(&self) -> &str { "" }
+impl TextureValues {
+    pub(super) fn new(context: &WebGlRenderingContext, index: u32, canvas: Canvas) -> anyhow::Result<TextureValues> {
+        let object = Texture::new();
+        let gl_value = Some(create_texture(context,(index,canvas))?);
+        Ok(TextureValues { gl_value, object })
+    }
 
-    fn activate(&self, context: &WebGlRenderingContext, gl_value: &(u32,WebGlTexture)) -> anyhow::Result<()> {
-        context.active_texture(WebGlRenderingContext::TEXTURE0+gl_value.0);
-        handle_context_errors(context)?;
-        context.bind_texture(WebGlRenderingContext::TEXTURE_2D,Some(&gl_value.1));
-        handle_context_errors(context)?;
+    pub(super) fn activate(&self, context: &WebGlRenderingContext) -> anyhow::Result<()> {
+        if let Some(buffer) = &self.gl_value {
+            context.active_texture(WebGlRenderingContext::TEXTURE0+buffer.0);
+            handle_context_errors(context)?;
+            context.bind_texture(WebGlRenderingContext::TEXTURE_2D,Some(&buffer.1));
+            handle_context_errors(context)?;
+        }
         Ok(())
     }
 
-    fn value_to_gl(&self, context: &WebGlRenderingContext, our_data: (u32,Canvas)) -> anyhow::Result<(u32,WebGlTexture)> {
-        let texture = context.create_texture().ok_or_else(|| err!("cannot create texture"))?;
-        handle_context_errors(context)?;
-        context.bind_texture(WebGlRenderingContext::TEXTURE_2D,Some(&texture));
-        handle_context_errors(context)?;
-        create_texture(context,our_data.1.element())?;
-        apply_weave(context,our_data.1.weave())?;
-        Ok((our_data.0,texture))
-    }
-
-    fn delete(&self, context: &WebGlRenderingContext, gl_value: &(u32,WebGlTexture)) -> anyhow::Result<()> {
-        context.delete_texture(Some(&gl_value.1));
-        handle_context_errors(context)?;
+    pub fn delete(&mut self, context: &WebGlRenderingContext) -> anyhow::Result<()> {
+        if let Some(gl_value) = &self.gl_value {
+            context.delete_texture(Some(&gl_value.1));
+            handle_context_errors(context)?;
+        }
         Ok(())
     }
 }
